added request handler for event controller

+++ back-end/src/api/controllers/event_controller.py
@@ -12,44 +12,73 @@ def get_events():
     """
         This is the endpoint returning event list
     """
-    api_response = event_service.get_events()
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        api_response = event_service.get_events()
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/<event_id>', methods=['GET'])
 def get_event(event_id):
     """
         This is the endpoint returning a single event with the given id
     """
-    api_response = event_service.get_event(event_id)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        api_response = event_service.get_event(event_id)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['POST'])
 def insert_event():
     """
         This is the endpoint for creating a new event
     """
+    req_handler_response = request_handler.validate_token(request)
 
-    data = request.get_json()
-    api_response = event_service.insert_event(data)
-    return api_response
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        data = request.get_json()
+        api_response = event_service.insert_event(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['DELETE'])
 def delete_event():
     """
         This is endpoint for deleting an event 
     """
-    data = request.get_json()
-    api_response = event_service.delete_event(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        data = request.get_json()
+        api_response = event_service.delete_event(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/', methods=['PUT'])
 def update_event():
     """
         This is endpoint for updating an event 
     """
-    data = request.get_json()
-    api_response = event_service.update_event(data)
-    return api_response
+    req_handler_response = request_handler.validate_token(request)
+
+    if req_handler_response['success']:
+        core_app_context = req_handler_response['core_app_context']
+        data = request.get_json()        
+        api_response = event_service.update_event(data)
+        return api_response
+
+    return req_handler_response
 
 @app.route(api_version + '/my-events', methods=['GET'])
 def my_events():
